I didn't go through everything but still :) more info in description
I think your code could use a single flow point.
you have handlers split on to different files.
and async accept any request and then call the correct handler for that type.
but that is a pitfall to many exploits and bugs.
*race conditions of different things resolving in different order due to client lag or double clicking.
*malicious user sending things out of order.
*no clear flow of logic.  instead it is pieced together like a graph xD ( one handler emitting events and so on ).  

---------------
I would not use this pattern of event emmitters and handlers but instead : 
have a central point of state management per client. ( I'm thinking of a client as a thread )
it has it's own flow of the game ( first i only accept init, than i accept placeShip, now i accept fire... )
handle requests with a switch case,
and dispatch to the correct handler to do the specific work.
that way you can ensure what events are acceptable at what state point.
without needing to cancel old event listeners and so on...
*this is not the use-case for this pattern. I think..